+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Added view count
     * Added publish date.
     * Made it so that after the view count it says Views (or view if it only has one view)
+    * Added better optimization (instead of `if, if else, if else, if else` instead it now uses switch and case)
 * Channel Views
     * Made it so if its over 1000 it turns into 1.0k, then if over 10,000 it will be 10k, and so on. It's limit is about 1 billion. (following + followers)
     * Optimized channel thumbnail
@@ -20,6 +21,7 @@
     * Made follower/following count bigger to fit results up to 4 digits
     * Made video thumbnails show gray box if loading
     * Fixed bug where anything over 1000 followers or following would show up with the .0k tag. (ex. 1032.2k)
+    * Added better optimization (instead of `if, if else, if else, if else` instead it now uses switch and case)
 * Camera Views
     * Going to camera will return it to the last position you left it at.
     * Fixed bug where first frame of video would be black making thumbnails completly black.